Refactor: Consolidate Firebase initialization and auth logic
This commit addresses issues with Firebase authentication by:

1.  **Consolidating Firebase Initialization:**
    *   Designated `react-pwa/src/firebase/config.js` as the single source for Firebase app initialization and service configuration.
    *   Merged necessary functionalities (like auth providers and helper functions) from `react-pwa/src/firebase.js` and `react-pwa/src/lib/firebase.js` into `config.js`.
    *   Removed the redundant `react-pwa/src/firebase.js` and `react-pwa/src/lib/firebase.js` files.

2.  **Consolidating Auth Context:**
    *   Standardized on `react-pwa/src/lib/auth-context.js` as the primary authentication context provider.
    *   Removed the simpler, redundant `AuthProvider` and `useAuth` hook from `react-pwa/src/hooks/useAuth.js`.
    *   Ensured the main `AuthProvider` correctly wraps the application.

3.  **Standardizing Firebase Service Imports:**
    *   Updated all imports throughout the `react-pwa/src/` directory to source Firebase services (auth, db, storage), SDK functions, and configured instances (like `googleProvider`) from the consolidated `react-pwa/src/firebase/config.js`. This ensures the entire application uses a single, consistently configured Firebase instance.

4.  **Verifying Google Sign-In:**
    *   Ensured that `react-pwa/src/hooks/useGoogleSignIn.js` uses the shared `googleProvider` instance from `config.js` rather than creating its own.

These changes eliminate potential conflicts arising from multiple Firebase initializations or auth contexts, leading to a more stable and predictable authentication system.
