+++ react-pwa/src/firebase/utils.js
@@ -1,23 +1,25 @@
-import { auth, db, storage } from './config';
 import { 
-  signInWithEmailAndPassword, 
-  createUserWithEmailAndPassword, 
-  signOut, 
+  auth, 
+  db, 
+  storage,
+  signInWithEmailAndPassword,
+  createUserWithEmailAndPassword,
+  signOut,
   sendPasswordResetEmail,
-  updateProfile
-} from 'firebase/auth';
-import { 
-  doc, 
-  setDoc, 
-  getDoc, 
-  updateDoc, 
-  deleteDoc, 
-  collection, 
-  query, 
+  updateProfile,
+  doc,
+  setDoc,
+  getDoc,
+  updateDoc,
+  deleteDoc,
+  collection,
+  query,
   getDocs,
-  serverTimestamp 
-} from 'firebase/firestore';
-import { ref, uploadBytes, getDownloadURL } from 'firebase/storage';
+  serverTimestamp,
+  ref,
+  uploadBytes,
+  getDownloadURL
+} from './config';
 
 // Auth functions
 export const signInWithEmail = (email, password) => {
