--- react-pwa/src/lib/firebase.js
@@ -1,76 +1,0 @@
-import { initializeApp, getApps, getApp } from 'firebase/app';
-import { 
-  getAuth, 
-  GoogleAuthProvider, 
-  signInWithEmailAndPassword, 
-  createUserWithEmailAndPassword, 
-  signInWithPopup, 
-  signOut, 
-  updateProfile,
-  onAuthStateChanged 
-} from 'firebase/auth';
-import { getFirestore, doc, getDoc, setDoc } from 'firebase/firestore';
-
-// Your web app's Firebase configuration
-const firebaseConfig = {
-  apiKey: process.env.REACT_APP_FIREBASE_API_KEY,
-  authDomain: process.env.REACT_APP_FIREBASE_AUTH_DOMAIN,
-  projectId: process.env.REACT_APP_FIREBASE_PROJECT_ID,
-  storageBucket: process.env.REACT_APP_FIREBASE_STORAGE_BUCKET,
-  messagingSenderId: process.env.REACT_APP_FIREBASE_MESSAGING_SENDER_ID,
-  appId: process.env.REACT_APP_FIREBASE_APP_ID,
-  measurementId: process.env.REACT_APP_FIREBASE_MEASUREMENT_ID
-};
-
-// Initialize Firebase
-let app;
-let auth;
-let db;
-let googleProvider;
-
-try {
-  // Check if Firebase app is already initialized
-  if (getApps().length === 0) {
-    // Initialize Firebase app if not already initialized
-    app = initializeApp(firebaseConfig);
-  } else {
-    // Use existing app if already initialized
-    app = getApp();
-  }
-  
-  // Initialize auth and firestore
-  auth = getAuth(app);
-  db = getFirestore(app);
-  googleProvider = new GoogleAuthProvider();
-  
-  // Add Google scopes if needed
-  googleProvider.addScope('profile');
-  googleProvider.addScope('email');
-  
-} catch (error) {
-  console.error('Firebase initialization error:', error);
-  // Fallback initialization in case of any errors
-  if (!app) {
-    app = initializeApp(firebaseConfig);
-    auth = getAuth(app);
-    db = getFirestore(app);
-    googleProvider = new GoogleAuthProvider();
-  }
-}
-
-export { 
-  auth, 
-  db, 
-  googleProvider,
-  signInWithEmailAndPassword,
-  createUserWithEmailAndPassword,
-  signInWithPopup,
-  signOut,
-  updateProfile,
-  onAuthStateChanged,
-  doc,
-  getDoc,
-  setDoc
-};
-
-export default app;
