+++ react-pwa/src/ui/dashboard/components/DashboardHeader.jsx.new
@@ -1,8 +1,7 @@
 import React, { useState } from 'react';
 import { Link, useNavigate } from 'react-router-dom';
 import { useAuth } from '../../../lib/auth-context';
-import { signOut } from 'firebase/auth';
-import { auth } from '../../../firebase/config';
+import { auth, signOut } from '../../../firebase/config';
 
 const DashboardHeader = ({ toggleSidebar }) => {
   const navigate = useNavigate();
